Revert "feat(tui): load external plugin exports"
This reverts commit fdd6498909c86cdd71d6cb30029ddc1d2928b06b.
